requirements.txt
# Core Dependencies
python-dotenv==1.0.1
numpy==2.2.4
pandas==2.2.3
scipy==1.15.2
scikit-learn==1.6.1
pydantic==2.10.6
pydantic-core==2.27.2
PyYAML==6.0.2
requests==2.32.3
tqdm==4.67.1
pillow==11.1.0
matplotlib==3.8.4
seaborn==0.13.2

# LangChain and Related
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-text-splitters==0.3.7
langchain-huggingface==0.1.2
langchain-openai==0.3.9
langchain-deepseek==0.1.2
langgraph==0.3.18
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.58
langsmith==0.3.18

# LlamaIndex and Related
llama-index==0.12.25
llama-index-core==0.12.25
llama-index-embeddings-huggingface==0.5.2
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.26
llama-index-readers-file==0.4.6
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1

# AI Models and APIs
openai==1.68.2
anthropic==0.49.0
cohere==5.14.0
transformers==4.49.0
sentence-transformers==3.4.1
huggingface-hub==0.29.3
tokenizers==0.21.1
tiktoken==0.9.0

# Vector Databases and Embeddings
faiss-cpu==1.10.0
chromadb==0.6.3

# Document Processing
pypdf==5.4.0
PyMuPDF==1.25.3
python-docx==1.1.2
python-pptx==1.0.2
beautifulsoup4==4.13.3
pdfplumber==0.11.5
pdf2image==1.17.0
pytesseract==0.3.13
camelot-py==1.0.0
unstructured==0.17.2

# HTTP and Networking
aiohttp==3.11.14
httpx==0.28.1
requests-toolbelt==1.0.0

# Data Formats
dataclasses-json==0.6.7
jsonpatch==1.33
jsonpointer==3.0.0
orjson==3.10.15

# Machine Learning and Deep Learning
torch==2.6.0
torchvision==0.21.0

# Utilities
tenacity==9.0.0
regex==2024.11.6
filelock==3.18.0
fsspec==2025.3.0
SQLAlchemy==2.0.39
Jinja2==3.1.6
sympy==1.13.1
typing-extensions==4.12.2
typing-inspect==0.9.0

# Development and Debugging
ipython==8.32.0
ipykernel==6.29.5
jupyter_client==8.6.3
jupyter_core==5.7.2
